requirements.txt
python-dotenv
azure-core
azure-identity
azure-search-documents
openai
pypdf
docx2txt
streamlit
streamlit-chat
langchain
langchain-core
langchain_openai
langchain-text-splitters
langchain-community
matplotlib
wordcloud
arxiv
semanticscholar